requirements.txt
alabaster==0.7.12
appnope==0.1.2
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
build==0.5.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
decorator==5.0.9
docutils==0.16
idna==2.10
imagesize==1.2.0
importlib-metadata==4.5.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
keyring==23.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
packaging==20.9
parso==0.8.2
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkginfo==1.7.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
pygame==2.0.1
Pygments==2.9.0
pymunk==6.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
readme-renderer==29.0
recommonmark==0.7.1
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.1.1
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxjp.themes.basicstrap==0.5.0
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
twine==3.4.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1